페이징 DB연결

+++ src/main/java/shop/mtcoding/blog/board/BoardRepository.java
@@ -60,11 +60,11 @@ public class BoardRepository {
 
     }
 
-    public BigInteger count(){
+    public int count() {
         Query query = em.createNativeQuery("select count(*) from board_tb");
-        BigInteger count = (BigInteger) query.getSingleResult();
-        return count ;
+
+        Long count = (Long) query.getSingleResult();
+            return count.intValue();
 
     }
-
 }
